# avoid excessive inlining of Ryu