--- a/base/ryu/shortest.jl
+++ b/base/ryu/shortest.jl
@@ -224,11 +224,10 @@ integer. If a `maxsignif` argument is provided, then `b < maxsignif`.
     return b, e10
 end
 
-
-@inline function writeshortest(buf::Vector{UInt8}, pos, x::T,
-                               plus=false, space=false, hash=true,
-                               precision=-1, expchar=UInt8('e'), padexp=false, decchar=UInt8('.'),
-                               typed=false, compact=false) where {T}
+function writeshortest(buf::Vector{UInt8}, pos, x::T,
+                       plus=false, space=false, hash=true,
+                       precision=-1, expchar=UInt8('e'), padexp=false, decchar=UInt8('.'),
+                       typed=false, compact=false) where {T}
     @assert 0 < pos <= length(buf)
     neg = signbit(x)
     # special cases