@oscardssmith 2021-04-24
What effect does this have on printing performance?