@bicycle1885 2021-04-24
This prints 10,000 floats generated from a standard normal distribution to /dev/null:
```julia
using BenchmarkTools

function benchmark(out, x)
    for xx in x
        println(out, xx)
    end
end

using Random
Random.seed!(1234)
x = randn(100_000)
@btime benchmark($devnull, x)
```

PR:
```
~/w/julia (ryu-noinline|…) $ ./julia test.jl
  15.545 ms (300000 allocations: 44.25 MiB)
~/w/julia (ryu-noinline|…) $ ./julia test.jl
  15.419 ms (300000 allocations: 44.25 MiB)
```

master:
```
~/w/julia (master|…) $ ./julia test.jl
  14.682 ms (300000 allocations: 44.25 MiB)
~/w/julia (master|…) $ ./julia test.jl
  15.021 ms (300000 allocations: 44.25 MiB)
~/w/julia (master|…) $ ./julia test.jl
  15.302 ms (300000 allocations: 44.25 MiB)
```